📦 NEW: Seperate routes from main app
