+++ src/App.jsx
@@ -1,19 +1,14 @@
-import { Suspense, lazy } from 'react'
+import { Suspense } from 'react'
 import { Helmet } from 'react-helmet-async'
 import { useSelector } from 'react-redux'
-import { Route, Switch } from 'react-router-dom'
 import { toast } from 'react-toastify'
 import { ThemeProvider } from 'styled-components'
 
 import { Header } from 'components/header'
 import { LoaderAnimation } from 'components/shared'
-import { PrivateRoute } from 'hoc'
+import { AppRoutes } from 'routes'
 import { selectTheme } from 'store/settingsSlice'
 import { DarkTheme, GlobalStyles, LightTheme } from 'styles'
-
-const Dashboard = lazy(() => import('pages/Dashboard'))
-const Login = lazy(() => import('pages/Login'))
-const NotFound = lazy(() => import('pages/NotFound'))
 
 const App = () => {
   toast.configure()
@@ -28,16 +23,11 @@ const App = () => {
           content="IIT Bombay's course resources sharing website"
         />
       </Helmet>
-      <Header />
-
       <GlobalStyles />
+
+      <Header />
       <Suspense fallback={<LoaderAnimation />}>
-        <Switch>
-          <Route path="/login" component={Login} />
-          <Route path="/404" component={NotFound} />
-          <PrivateRoute path="/" component={Dashboard} />
-          <Route component={NotFound} />
-        </Switch>
+        <AppRoutes />
       </Suspense>
     </ThemeProvider>
   )
